+++ tests/unit/test_grpc_helpers.py
@@ -12,9 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import grpc
 import mock
 import pytest
+
+try:
+    import grpc
+except ImportError:
+    pytest.skip("No GRPC", allow_module_level=True)
 
 from google.api_core import exceptions
 from google.api_core import grpc_helpers
@@ -50,6 +54,9 @@ class RpcErrorImpl(grpc.RpcError, grpc.Call):
         return self._code
 
     def details(self):
+        return None
+
+    def trailing_metadata(self):
         return None